==> notebooks/validation/01 Fetch Wind Toolkit Data.ipynb <==
## WTK Data Preparation

Caleb Phillips (caleb.phillips@nrel.gov), Dmitry Duplyakin (dmitry.duplyakin@nrel.gov) and Jenna Ruzekowicz (jenna.ruzekowicz@nrel.gov)

The purpose of this notebook is to read in WTK and WTK-LED data at the turbine locations (the entire time period available, spatially and vertically interpolated).

Notes: 
Might need to install Rex if it isn't installed already:
conda install nrel-rex --channel=nrel

More about rex: https://github.com/NREL/rex
2018 5-min monthly h5 (the file you referenced on the 21st):
/campaign/tap/CONUS/wtk/5min/2018/{month}/conus_2018-{month}.h5
 
2018 5-min yearly h5 slices:
/shared-projects/wtk-led/CONUS/wtk/2018/yearly_h5/conus_2018_{height}m.h5
 
2019 60-min yearly h5:
/campaign/tap/CONUS/wtk/60min/2019/conus_2019.h5

import numpy as np
import pandas as pd
import geopandas as gpd

import h5pyd
from rex.resource_extraction import MultiYearWindX
from dw_tap.data_fetching import get_data_wtk_led_on_eagle 
from dw_tap.data_fetching import getData

from tqdm import tqdm

from site_index import SiteIndex

#Added for OneEnergy
from site_index_oe import SiteIndexOE

### Read in index of turbines

If running just the windtoolkit code, you only need this not the next two sections.

#Bergey
#index = SiteIndex()

#One Energy
index = SiteIndexOE()

index.head()

### Fetch Wind Toolkit Data

# Open the wind data "file"
# server endpoint, username, password are found in ~/.hscfg
f = h5pyd.File("/nrel/wtk-us.h5", 'r', bucket="nrel-pds-hsds") 

def fetch_dfs(index,wtk_dfs=[]):
    for row in tqdm(index.itertuples()):
        tid = row[1]
        lat = row[6]
        lon = row[7]
        z_turbine = row[8]
        # avoid re-running things that have already run
        if tid in [x['tid'][0] for x in wtk_dfs]:
            print("Skipping "+str(tid))
            continue
        atmospheric_df = getData(f, lat, lon, z_turbine, "IDW", 
                                 power_estimate=True,
                                 inverse_monin_obukhov_length=True)
        atmospheric_df['tid'] = tid
        wtk_dfs.append(atmospheric_df)

    return wtk_dfs

# in case HSDS has a connection error we will try a second time
try:
    wtk_dfs = fetch_dfs(index)
except ConnectionError:
    print("Caught a connection error, trying to resume...")
    wtk_dfs = fetch_dfs(indx,wtk_dfs)

wtk_df = pd.concat(wtk_dfs)

wtk_df.head()

wtk_df['packet_date'] = pd.to_datetime(wtk_df["datetime"]).dt.tz_localize('UTC')

#Bergey
#wtk_df.to_csv("01 Bergey Turbine Data/wtk_tp.csv.bz2",index=False)

#One Energy
wtk_df.to_csv("01 One Energy Turbine Data/wtk_tp.csv.bz2",index=False)

### Fetch WTK-LED Data

**Requirement:** The following code must run on `Eagle` to access h5 files for WTK-LED

#### 2018 5-minute data from WTK-LED

#Bergy
#index = pd.read_csv("01 Bergey Turbine Data/bergey_sites.csv")

#One Energy
index = pd.read_csv("01 One Energy Turbine Data/OneEnergyTurbineData.csv")

index.head()

def fetch_dfs_wtk_led_2018(index, wtk_dfs=[]):
    # 12 monthly files for 2018
    files = ['/campaign/tap/CONUS/wtk/5min/2018/%s/conus_2018-%s.h5' % \
             (str(i).zfill(2), str(i).zfill(2)) for i in range(1,13)]
    
    for row in tqdm(index.itertuples()):
        tid = row[1]
        lat = row[5]
        lon = row[6]
        z_turbine = row[7]
        # avoid re-running things that have already run
        if tid in [x['tid'][0] for x in wtk_dfs]:
            print("Skipping "+str(tid))
            continue
            
        atmospheric_df = pd.DataFrame()
        # Iterate over all monthly files
        for file in files:
            myr = MultiYearWindX(file, hsds=False)
            d = get_data_wtk_led_on_eagle(myr, 
                                          lat, lon, z_turbine, "IDW", 
                                          power_estimate=False,
                                          start_time_idx=None, 
                                          end_time_idx=None,
                                          time_stride=None)
            atmospheric_df = pd.concat([atmospheric_df, d])
        
        atmospheric_df['tid'] = tid
        atmospheric_df.reset_index(drop=True, inplace=True)
        wtk_dfs.append(atmospheric_df)

    return wtk_dfs

# in case HSDS has a connection error we will try a second time
try:
    wtk_dfs = fetch_dfs_wtk_led_2018(index)
except ConnectionError:
    print("Caught a connection error, trying to resume...")
    wtk_dfs = fetch_dfs_wtk_led_2018(indx,wtk_dfs)

wtk_df_2018 = pd.concat(wtk_dfs)
wtk_df_2018.head()

len(wtk_df_2018)

wtk_df_2018.tid.value_counts()

# The following causes the error: Already tz-aware, use tz_convert to convert.
#wtk_df_2018['packet_date'] = \
#    pd.to_datetime(wtk_df_2018["datetime"]).dt.tz_localize('UTC')
# Skip localization: 
wtk_df_2018['packet_date'] = wtk_df_2018["datetime"]

#Bergey
#wtk_df_2018[['tid','packet_date','ws','wd']].\
#    to_csv("01 Bergey Turbine Data/wtk_led_2018.csv.bz2",index=False)

#One Energy
wtk_df_2018[['tid','packet_date','ws','wd']].\
    to_csv("01 One Energy Turbine Data/wtk_led_2018.csv.bz2",index=False)

#### 2019 hourly data from WTK-LED

def fetch_dfs_wtk_led_2019(index, wtk_dfs=[]):

    # 2019 hourly file
    myr = MultiYearWindX('/campaign/tap/CONUS/wtk/60min/2019/conus_2019.h5', hsds=False)
    
    for row in tqdm(index.itertuples()):
        tid = row[1]
        lat = row[5]
        lon = row[6]
        z_turbine = row[7]
        # avoid re-running things that have already run
        if tid in [x['tid'][0] for x in wtk_dfs]:
            print("Skipping "+str(tid))
            continue
            
        atmospheric_df = pd.DataFrame()

        atmospheric_df = get_data_wtk_led_on_eagle(myr, 
                                          lat, lon, z_turbine, "IDW", 
                                          power_estimate=False,
                                          start_time_idx=None, 
                                          end_time_idx=None,
                                          time_stride=None)
        
        atmospheric_df['tid'] = tid
        wtk_dfs.append(atmospheric_df)

    return wtk_dfs

# in case HSDS has a connection error we will try a second time
try:
    wtk_dfs = fetch_dfs_wtk_led_2019(index)
except ConnectionError:
    print("Caught a connection error, trying to resume...")
    wtk_dfs = fetch_dfs_wtk_led_2019(indx,wtk_dfs)

wtk_df_2019 = pd.concat(wtk_dfs)
wtk_df_2019.head()

len(wtk_df_2019)

# The following causes the error: Already tz-aware, use tz_convert to convert.
#wtk_df_2019['packet_date'] = \
#    pd.to_datetime(wtk_df_2019["datetime"]).dt.tz_localize('UTC')
# Skip localization: 
wtk_df_2019['packet_date'] = wtk_df_2019["datetime"]

#Bergey
#wtk_df_2019[['tid','packet_date','ws','wd']].\
#    to_csv("01 Bergey Turbine Data/wtk_led_2019.csv.bz2",index=False)

#One Energy
wtk_df_2019[['tid','packet_date','ws','wd']].\
    to_csv("01 One Energy Turbine Data/wtk_led_2019.csv.bz2",index=False)

wtk_df_2019.tid.value_counts()
