==> notebooks/validation/01 Read in Data.ipynb <==
Jenna Ruzekowicz (jenna.ruzekowicz@nrel.gov) and Caleb Phillips (caleb.phillips@nrel.gov)

The purpose of this notebook is to read in two sets of data over the same time period for comparison:
1) WTK-LED data
2) Either: 
    a) power output and/or wind speed data from turbine(s) 
    b) wind speed measurements from met tower(s)

Data sets are matched on location and time stamp.

A combined and labeled dataframe will be exported to a csv file in the "01Data" folder. The naming convention for the csv file will be as follows: source_lat_lon_startdate_enddate.csv
where source is either "bergey", "oneenergy"...

Notes: 
- really long string potential issue
- simple naming, numbering? 
- number cases with file as reference sheet
- metadata global view dataframe
- lat lon heights stag of analysis
- what's processed and what is not? 

Notes: 
Might need to install Rex if it isn't installed already:
conda install nrel-rex --channel=nrel

More about rex: https://github.com/NREL/rex
2018 5-min monthly h5 (the file you referenced on the 21st):
/campaign/tap/CONUS/wtk/5min/2018/{month}/conus_2018-{month}.h5
 
2018 5-min yearly h5 slices:
/shared-projects/wtk-led/CONUS/wtk/2018/yearly_h5/conus_2018_{height}m.h5
 
2019 60-min yearly h5:
/campaign/tap/CONUS/wtk/60min/2019/conus_2019.h5

import numpy as np
import pandas as pd
import geopandas as gpd
from rex.resource_extraction import MultiYearWindX
from dw_tap.data_fetching import get_data_wtk_led_on_eagle

Step 1) Read in either power output/wind speed data for wind turbine(s) or wind speed data from met tower(s)

#Reading in data from W1 turbine at Marion OH location (oneenergy turbine), 2018
power_output_df = pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180131.000000.marion.w1.xlsx", header=0, usecols="B, C, M")
power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180228.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180331.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180430.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180531.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180630.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180731.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180831.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20180930.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20181031.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20181130.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
#power_output_df = pd.concat([power_output_df, pd.read_excel("../../data/marionw1/turbine.oneenergy.00.20181231.000000.marion.w1.xlsx", header=0, usecols="B, C, M")])
power_output_df.rename(columns={'Time':'datetime', 'WindTurbineEnergyYield_kWh_':'measured_production', 'AvgWindSpeed_m_s_':'measured_ws'}, inplace=True)
#Below line is required time conversion, only run once
power_output_df['datetime'] = power_output_df['datetime'].dt.tz_localize('UTC')

print(power_output_df)

power_output_df.info()

Step 2) Read in WTK-LED data for 2018 at the same location

z_turbine = 80
lat, lon = 40.591555, -83.182092
#change 13 for 3 here until I have accesss to all data
files = ['/campaign/tap/CONUS/wtk/5min/2018/%s/conus_2018-%s.h5' % (str(i).zfill(2), str(i).zfill(2)) for i in range(1,3)]

atmospheric_df = pd.DataFrame()
for file in files:

    myr = MultiYearWindX(file, hsds=False)

    d = get_data_wtk_led_on_eagle(myr, 
                                  lat, lon, z_turbine, "IDW", 
                                  power_estimate=False,
                                  start_time_idx=None, 
                                  end_time_idx=None,
                                  time_stride=None)
    atmospheric_df = pd.concat([atmospheric_df, d])
print(atmospheric_df)

atmospheric_df.info()

Step 3) Merge the two dataframes

analysis_df = atmospheric_df.merge(power_output_df[['datetime', 'measured_production', 'measured_ws']], on='datetime', how='left')

analysis_df = analysis_df.

analysis_df[~analysis_df['measured_ws'].isna()]

Notes: 
Lots of different inputs
naming convention
file checking
function in package
check validation case (function) -> gets path dir, checks for all variables? 
does it have atmo. obstacles. 
checker within the for loop

before we read the data
what do we do with no obstacles? 

in each dir (each analysis case)-> lat, lon, and height
some type of checking

tangent: 
notebooks to always save intermediate results! 
have a naming convention 
organize the intermediate steps
