==> notebooks/validation/04 Compute Daily Metrics.ipynb <==
## Performance Analysis for Daily Energy Production

Caleb Phillips (caleb.phillips@nrel.gov)

The purpose of this notebook is to read in computed predictions and compare them to actual observations, computing metrics. 

from common import *
import pandas as pd
import numpy as np
from tqdm import tqdm
import glob
import re
import os.path
import importlib
from dw_tap import power_output
import site_index
import plotly.express as px
plotly.offline.init_notebook_mode()

index = site_index.SiteIndex()

### Load Ground Truth (Bergey Daily Data)

daily_bergey = pd.read_csv("01 Bergey Turbine Data/daily_summaries.csv.bz2")
daily_bergey = daily_bergey.merge(index.index[['APRS ID','AID']].rename(columns={'APRS ID':'tid'}))
daily_bergey['date'] = pd.to_datetime(daily_bergey['date'],utc=True).dt.date
daily_bergey.head()

#### Evaluate faults

daily_bergey["fault"] = "None"
daily_bergey.loc[daily_bergey['faults'].str.contains("FAULT"),"fault"] = "Fault"
daily_bergey.loc[daily_bergey['faults'].str.contains("WAITING INITIALIZING"),"fault"] = "Reset"
daily_bergey.loc[daily_bergey['faults'].str.contains("MANUAL STOP"),"fault"] = "Stopped"
daily_bergey.loc[daily_bergey['soft_grid'].str.contains("%"),"fault"] = "Curtailed"
daily_bergey["fault"].value_counts()

### Load Prediction Data for all Models/Sites

providers = ["bergey"]
models = ["perera","anl","lanl"]
wind_sources = ["wtk","wtk_bc","wtk_led_2018","wtk_led_2019","wtk_led_bc"]
obs_groups = ["all","bldgsonly","treesasbldgs","bldgsonly_100m"]

dfs = []

for provider in providers:
    for model in models:
        for wind_source in wind_sources:
            for og in obs_groups:
                for tid in index.tids(True):
                    fname = f"03 Model Outputs/{provider}_{model}_{tid}_{wind_source}_{og}.csv.bz2"
                
                    if not os.path.exists(fname):
                        continue

                    print(fname)

                    d = pd.read_csv(fname)
                    d["model"] = model
                    d["wind_source"] = wind_source
                    d["provider"] = provider
                    d["tid"] = tid
                    d["obsgroup"] = og
                    dfs.append(d[["model","obsgroup","wind_source","provider","tid","datetime","ws-adjusted"]])

                    if model == "perera":
                        d2 = d.copy()
                        d2["ws-adjusted"] = d2["ws-adjusted-2"]
                        d2["model"] = "shelter"
                        d2["obsgroup"] = og
                        dfs.append(d2[["model","obsgroup","wind_source","provider","tid","datetime","ws-adjusted"]])

                        d3 = d.copy()
                        d3["ws-adjusted"] = d3["ws-adjusted-3"]
                        d3["model"] = "shelter+"
                        d3["obsgroup"] = og
                        dfs.append(d3[["model","obsgroup","wind_source","provider","tid","datetime","ws-adjusted"]])

# No model data
for provider in providers:
    for wind_source in tqdm(wind_sources):
        if wind_source == "wtk":
            d = pd.read_csv("01 Bergey Turbine Data/wtk.csv.bz2")
            d.rename(columns={"ws":"ws-adjusted"},inplace=True)
            
        elif wind_source == "wtk_led_2018":
            d = pd.read_csv("01 Bergey Turbine Data/wtk_led_2018.csv.bz2")
            d.rename(columns={"ws":"ws-adjusted","packet_date":"datetime"},inplace=True)
            
        elif wind_source == "wtk_led_2019":
            d = pd.read_csv("01 Bergey Turbine Data/wtk_led_2019.csv.bz2")
            d.rename(columns={"ws":"ws-adjusted","packet_date":"datetime"},inplace=True)
            
        elif wind_source == "wtk_bc":
            d = pd.read_csv("02 Bias Correction/wtk_bc.csv.bz2")
            d.rename(columns={"ws_bc":"ws-adjusted"},inplace=True)
            
        elif wind_source == "wtk_led_bc":
            d = pd.read_csv("02 Bias Correction/wtk_led_bc.csv.bz2")
            d.rename(columns={"ws_bc":"ws-adjusted"},inplace=True)
            
        else:
            print("Unsupported wind source")
            
        d["wind_source"] = wind_source
        d["model"] = "none"
        d["provider"] = provider
        d["obsgroup"] = "none"
        dfs.append(d[["model","obsgroup","wind_source","provider","tid","datetime","ws-adjusted"]])

bigdf = pd.concat(dfs)
bigdf.head()

# filter to just those with lidar data
bigdf = bigdf.loc[bigdf["tid"].isin(index.tids(True))]

bigdf["power_kw"] = power_output.Bergey10.windspeed_to_kw(bigdf,'ws-adjusted')
bigdf['datetime'] = pd.to_datetime(bigdf['datetime'],format="%Y-%m-%d %H:%M:%S",utc=True) # this is slow
bigdf['date'] = bigdf['datetime'].dt.date
bigdf.head()

daily_summaries = bigdf[['model','obsgroup','wind_source','provider','tid','date','power_kw','ws-adjusted']]\
    .groupby(['model','obsgroup','wind_source','provider','tid','date'])\
    .agg({ 'power_kw': ['count','sum'],'ws-adjusted': ['mean','max','min','median'] })
daily_summaries.columns = ['_'.join(col) for col in daily_summaries.columns.values]
daily_summaries = daily_summaries.reset_index()
# scale the 5 min data so that it is kwh (too)
daily_summaries["power_kw_sum"] = daily_summaries["power_kw_sum"]/(daily_summaries["power_kw_count"]/24)
daily_summaries.rename(columns={"power_kw_sum":"energy_estimate_kwh",
                                "ws-adjusted_mean":"windspeed_estimate_mps_mean",
                                "ws-adjusted_median":"windspeed_estimate_mps_median",
                                "ws-adjusted_min":"windspeed_estimate_mps_min",
                                "ws-adjusted_max":"windspeed_estimate_mps_max"},inplace=True)
del daily_summaries['power_kw_count'] # no longer needed
daily_summaries.head()

### Create Merged Dataframe

merged_df = daily_summaries.merge(daily_bergey[['date','tid','energy_kwh','fault']],on=['tid','date'],how='right')
merged_df = merged_df.dropna().drop_duplicates() # omit days where we are missing data and duplicate rows from merge
merged_df.rename(columns={"energy_kwh":"energy_observed_kwh"},inplace=True)
merged_df.head()

px.histogram(daily_bergey,x='energy_kwh')

daily_bergey['energy_kwh'].quantile(0.90)

# error percent compared to actual generation
merged_df["relative_error_pct"] = ((merged_df["energy_estimate_kwh"] - merged_df["energy_observed_kwh"])/merged_df["energy_observed_kwh"])*100 
# error percent of max generation (ie 300kwh)
merged_df["relative_error_pct_of_max"] = ((merged_df["energy_estimate_kwh"] - merged_df["energy_observed_kwh"])/300)*100 

merged_df['date'] = pd.to_datetime(merged_df['date'])
merged_df['month'] = merged_df['date'].dt.month
merged_df['year'] = merged_df['date'].dt.year
merged_df['moyr'] = merged_df['date'].dt.year.astype(str)+"-"+merged_df['date'].dt.month.astype(str)

### Performance Plots: Overall

filtered_merged_df = merged_df[(merged_df["fault"] == "None")]

fig = px.scatter(filtered_merged_df[filtered_merged_df.obsgroup.isin(["treesasbldgs","none"])],
                 x="energy_observed_kwh",y="energy_estimate_kwh",
                 facet_row="model",facet_col="wind_source",
                 labels={"energy_estimate_kwh":"Pred(kwh)","energy_observed_kwh":"Obs(kwh)"})
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.scatter(filtered_merged_df[filtered_merged_df.obsgroup.isin(["treesasbldgs","none"])],
                 x="energy_observed_kwh",y="energy_estimate_kwh",
                 facet_row="model",facet_col="wind_source",color="fault",
                 labels={"energy_estimate_kwh":"Pred(kwh)","energy_observed_kwh":"Obs(kwh)"})
fig.update_yaxes(matches=None)
fig.update_xaxes(matches=None)
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.density_contour(filtered_merged_df[filtered_merged_df.obsgroup.isin(["treesasbldgs","none"])],
                         x="energy_observed_kwh",y="energy_estimate_kwh",
                         facet_row="model",facet_col="wind_source",
                         labels={"energy_estimate_kwh":"Pred(kwh)","energy_observed_kwh":"Obs(kwh)"})
fig.update_xaxes(range=[0, 40])
fig.update_yaxes(range=[0, 40])
fig.update_traces(contours_coloring="fill", contours_showlabels = True)
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.box(filtered_merged_df,x="model",y="relative_error_pct_of_max",color="wind_source",
             labels={"relative_error_pct_of_max":"Relative Error (%)","model":"Model"},
             points=False,title="Daily Relative Error (% of Max Production)")
fig.show()

fig = px.box(filtered_merged_df,x="wind_source",y="relative_error_pct_of_max",color="model",
             labels={"relative_error_pct_of_max":"Relative Error (%)","wind_source":"Wind Source"},
             points=False,title="Daily Relative Error (% of Max Production)")
fig.show()

# Defining function for 50th Percentile
p25 = lambda x: x.quantile(0.25)
p75 = lambda x: x.quantile(0.75)
safemean = lambda x: np.mean([0 if np.isnan(y) else y for y in x])

daily_perf_summary = filtered_merged_df.groupby(["model","obsgroup","wind_source"]).\
    agg({ "relative_error_pct_of_max": ["mean","median",p25,p75] }).\
    rename(columns={"<lambda_0>":'25%ile',"<lambda_1>":'75%ile'})

#daily_perf_summary.loc[["lanl","none"]]
daily_perf_summary

daily_perf_summary.columns = ['_'.join(col) for col in daily_perf_summary.columns.values]
daily_perf_summary.reset_index(inplace=True)
daily_perf_summary.to_csv("04 Error Metrics/bergey_daily_relative_error.csv.bz2",index=False)

#### Daily Results By Site

daily_bysite = filtered_merged_df.groupby(["model","obsgroup","wind_source","tid"]).\
    agg({ "relative_error_pct_of_max": ["mean","median",p25,p75],
          "energy_observed_kwh": ["mean","median",p25,p75],
          "energy_estimate_kwh": ["mean","median",p25,p75]}).\
    rename(columns={"<lambda_0>":'25%ile',"<lambda_1>":'75%ile'})
daily_bysite.columns = ['_'.join(col) for col in daily_bysite.columns.values]
daily_bysite.reset_index(inplace=True)
daily_bysite.head()

daily_bysite.to_csv("04 Error Metrics/bergey_daily_relative_error_by_site.csv.bz2",index=False)

fig = px.bar(daily_bysite[daily_bysite['wind_source'].isin(["wtk","wtk_bc"])],x="tid",
             y="relative_error_pct_of_max_mean",color="model",facet_col="wind_source",
             facet_row="obsgroup",barmode='group',
             labels={"relative_error_pct_of_max_mean":"Relative Error (%)",
                     "tid":"Site","model":"Model"},
             title="Mean Relative Error in Daily Energy Estimate (% of kwh) by Model/Site")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.bar(daily_bysite[daily_bysite['wind_source'].isin(["wtk"]) & daily_bysite['model'].isin(["anl","none"])],x="tid",
             y="relative_error_pct_of_max_mean",color="obsgroup",barmode='group',
             labels={"relative_error_pct_of_max_mean":"Relative Error (%)",
                     "tid":"Site","model":"Model","obsgroup":"Obstacles"},
             title="Mean Relative Error in Daily Energy Estimate (% of kwh) by Model/Site - ANL vs None")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.bar(daily_bysite[daily_bysite['wind_source'].isin(["wtk_bc"]) & daily_bysite['model'].isin(["anl","none"])],x="tid",
             y="relative_error_pct_of_max_mean",color="obsgroup",barmode='group',
             labels={"relative_error_pct_of_max_mean":"Relative Error (%)",
                     "tid":"Site","model":"Model","obsgroup":"Obstacles"},
             title="Mean Relative Error in Daily Energy Estimate (% of kwh) by Model/Site - ANL vs None")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

    fig = px.bar(daily_bysite[daily_bysite['wind_source'].isin(["wtk"]) & daily_bysite['model'].isin(["lanl","none"])],x="tid",
                 y="relative_error_pct_of_max_mean",color="obsgroup",barmode='group',
                 labels={"relative_error_pct_of_max_mean":"Relative Error (%)",
                         "tid":"Site","model":"Model","obsgroup":"Obstacles"},
                 title="Mean Relative Error in Daily Energy Estimate (% of kwh) by Model/Site - LANL vs WTK")
    fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
    fig.show()

fig = px.bar(daily_bysite[daily_bysite['wind_source'].isin(["wtk_bc"]) & daily_bysite['model'].isin(["lanl","none"])],x="tid",
             y="relative_error_pct_of_max_mean",color="obsgroup",barmode='group',
             labels={"relative_error_pct_of_max_mean":"Relative Error (%)",
                     "tid":"Site","model":"Model","obsgroup":"Obstacles"},
             title="Mean Relative Error in Daily Energy Estimate (% of kwh) by Model/Site - LANL vs WTK+BC")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.box(filtered_merged_df[filtered_merged_df['wind_source'].isin(["wtk","wtk_bc"])],x="tid",
             y="relative_error_pct_of_max",color="model",facet_row="wind_source",facet_col="obsgroup",
             labels={"relative_error_pct_of_max":"Relative Error (%)",
                     "tid":"Site","model":"Model"},points=False,
             title="Mean Relative Error in Daily Energy Estimate (% of kwh) by Model/Site")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.violin(filtered_merged_df[filtered_merged_df['wind_source'].isin(["wtk"]) & filtered_merged_df['model'].isin(["anl","none"])],
             y="tid",color='obsgroup',
             x="relative_error_pct_of_max",facet_row="wind_source",
             labels={"relative_error_pct_of_max":"Relative Error (%)",
                     "tid":"Site","obsgroup":"Obstacles"},points=False,
             width=700,height=1000,
             title="Mean Relative Error in Daily Energy Estimate (% of kwh) by Model/Site (WTK Wind Source)")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.histogram(filtered_merged_df[filtered_merged_df['wind_source'].isin(["wtk"]) & filtered_merged_df['model'].isin(["anl","none"])],
             x="energy_estimate_kwh",color="obsgroup",facet_col="tid",
             labels={"energy_estimate_kwh":"Energy (kwh)",
                     "tid":"Site","obsgroup":"Obstacles"},facet_col_wrap=4)
fig.update_yaxes(matches=None)
# Overlay both histograms
fig.update_layout(barmode='overlay')
# Reduce opacity to see both histograms
fig.update_traces(opacity=0.75)
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))

fig.show()

fig = px.box(filtered_merged_df[filtered_merged_df['wind_source'].isin(["wtk"]) & filtered_merged_df['model'].isin(["anl","none"])],
             x="energy_estimate_kwh",color="obsgroup",y="tid",
             labels={"relative_error_pct_of_max":"Relative Error (%)",
                     "tid":"Site","obsgroup":"Obstacles",
                     "energy_estimate_kwh":"Daily Energy Estimate (kwh)"},height=750,points=False)
fig.show()

#### Overall Results by Month and Site

# Note this includes days with faults, so curtailment, etc. is included
bysitebymonth = merged_df.groupby(["model","obsgroup","wind_source","tid","moyr","month"]).\
    agg({ "energy_estimate_kwh": "sum", "energy_observed_kwh": "sum" })
#bysitebymonth.columns = ['_'.join(col) for col in bysitebymonth.columns.values]
bysitebymonth.reset_index(inplace=True)
# Using 300kwh * 30 days for maximum production in calculation of relative error
bysitebymonth["relative_error_pct_of_max"] = ((bysitebymonth["energy_estimate_kwh"] - bysitebymonth["energy_observed_kwh"])/(300*30))*100 
bysitebymonth["relative_error_pct"] = ((bysitebymonth["energy_estimate_kwh"] - bysitebymonth["energy_observed_kwh"])/bysitebymonth["energy_observed_kwh"])*100 
bysitebymonth

bysitebymonth.to_csv("04 Error Metrics/bergey_monthly_relative_error_by_site.csv.bz2",index=False)

fig = px.box(bysitebymonth[bysitebymonth['wind_source'].isin(["wtk","wtk_bc"]) & bysitebymonth['model'].isin(["anl","none"])],x="tid",
             y="relative_error_pct_of_max",color="obsgroup",facet_row="wind_source",
             labels={"relative_error_pct_of_max":"Relative Error (%)",
                     "tid":"Site","obsgroup":"Obstacles"})
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.box(bysitebymonth[bysitebymonth['wind_source'].isin(["wtk","wtk_bc"]) & bysitebymonth['model'].isin(["perera","none"])],x="tid",
             y="relative_error_pct_of_max",color="obsgroup",facet_row="wind_source",
             labels={"relative_error_pct_of_max":"Relative Error (%)",
                     "tid":"Site","obsgroup":"Obstacles"})
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

p25 = lambda x: x.quantile(0.25)
p75 = lambda x: x.quantile(0.75)
safemean = lambda x: np.mean([0 if np.isnan(y) else y for y in x])

monthly_perf_summary = bysitebymonth[['model','wind_source','obsgroup','relative_error_pct_of_max']].\
    groupby(['model','wind_source','obsgroup']).\
    agg({ "relative_error_pct_of_max": ["mean","median",p25,p75]}).\
    rename(columns={"<lambda_0>":'25%ile',"<lambda_1>":'75%ile'})

monthly_perf_summary

monthly_perf_summary.columns = ['_'.join(col) for col in monthly_perf_summary.columns.values]
monthly_perf_summary.reset_index(inplace=True)
monthly_perf_summary.to_csv("04 Error Metrics/bergey_monthly_relative_error.csv.bz2",index=False)

#### Overall Results by Year

# Note this includes days with faults, so curtailment, etc. is included
bysitebyyear = merged_df.groupby(["model","obsgroup","wind_source","tid","year"]).\
    agg({ "energy_estimate_kwh": ["sum","count"], "energy_observed_kwh": "sum" })
bysitebyyear.columns = ['_'.join(col) for col in bysitebyyear.columns.values]
bysitebyyear.reset_index(inplace=True)
bysitebyyear.rename(columns={"energy_estimate_kwh_sum":"energy_estimate_kwh","energy_estimate_kwh_count":"count","energy_observed_kwh_sum":"energy_observed_kwh"},inplace=True)
bysitebyyear
# Using 300kwh * 365 days for maximum production in calculation of relative error
bysitebyyear["relative_error_pct_of_max"] = ((bysitebyyear["energy_estimate_kwh"] - bysitebyyear["energy_observed_kwh"])/(300*365))*100 
bysitebyyear["relative_error_pct"] = ((bysitebyyear["energy_estimate_kwh"] - bysitebyyear["energy_observed_kwh"])/bysitebyyear["energy_observed_kwh"])*100 

dc = index.index[['APRS ID','Bergey Generation (kWh) (0% Loss Assumption)']].rename(columns={"APRS ID":"tid","Bergey Generation (kWh) (0% Loss Assumption)":"bergey_kwh"})
bysitebyyear = bysitebyyear.merge(dc,on="tid")
bysitebyyear["bergey_relative_error_pct_of_max"] = ((bysitebyyear["bergey_kwh"] - bysitebyyear["energy_observed_kwh"])/(300*365))*100 
bysitebyyear.rename(columns={"bergey_kwh":"energy_bergey_estimate_kwh"},inplace=True)
bysitebyyear = bysitebyyear[bysitebyyear['count'] > 300]
bysitebyyear

bysitebyyear.to_csv("04 Error Metrics/bergey_annual_relative_error_by_site.csv.bz2",index=False)

puddle = bysitebyyear.where(bysitebyyear['count']>360)[['model','obsgroup','wind_source','tid','year','energy_estimate_kwh','energy_observed_kwh','energy_bergey_estimate_kwh']].\
    melt(id_vars=['model','obsgroup','wind_source','tid','year'])
fig = px.box(puddle.where((puddle.obsgroup=="bldgsonly") & (puddle.wind_source=='wtk')).dropna(),y="value",x="tid",color="variable",facet_row="model",facet_col="obsgroup",labels={"value":"Production (kwh)"})
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.box(puddle.where((puddle.obsgroup=="bldgsonly") & (puddle.wind_source=='wtk_bc')).dropna(),y="value",x="tid",color="variable",facet_row="model",facet_col="obsgroup",labels={"value":"Production (kwh)"})
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.box(puddle.where((puddle.obsgroup=="none") & (puddle.wind_source.isin(['wtk','wtk_bc']))).dropna(),y="value",x="tid",color="variable",facet_row="wind_source",labels={"value":"Production (kwh)"})
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.box(bysitebyyear,y="relative_error_pct_of_max",x="wind_source",color="model",facet_row="obsgroup",
            labels={"relative_error_pct_of_max":"Error(%)","wind_source":"Wind Source","model":"Model"})
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

fig = px.box(bysitebyyear[(bysitebyyear["wind_source"].isin(["wtk","wtk_bc"]))],y="relative_error_pct_of_max",x="wind_source",color="model",
       labels={"wind_source":"Wind Source","pct": "Annualized Error Percentage (%)","model":"Model",
              "relative_error_pct_of_max":"Error (%)"},
      facet_row="obsgroup")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

p25 = lambda x: x.quantile(0.25)
p75 = lambda x: x.quantile(0.75)
safemean = lambda x: np.mean([0 if np.isnan(y) else y for y in x])

yearly_perf_summary = bysitebyyear[['model','wind_source','obsgroup','relative_error_pct_of_max','bergey_relative_error_pct_of_max']].groupby(['model','wind_source','obsgroup']).\
    agg({ "relative_error_pct_of_max": ["mean","median",p25,p75] }).\
    rename(columns={"<lambda_0>":'25%ile',"<lambda_1>":'75%ile'})
#yearly_perf_summary
yearly_perf_summary.loc[["lanl","none"]]

yearly_perf_summary.columns = ['_'.join(col) for col in yearly_perf_summary.columns.values]
yearly_perf_summary.reset_index(inplace=True)
yearly_perf_summary.to_csv("04 Error Metrics/bergey_annual_relative_error.csv.bz2",index=False)

yearly_perf_summary_fullyears = bysitebyyear.where(bysitebyyear['count']>360)[['model','wind_source','obsgroup','relative_error_pct_of_max','bergey_relative_error_pct_of_max']].groupby(['model','wind_source','obsgroup']).\
    agg({ "relative_error_pct_of_max": ["mean","median",p25,p75], 'bergey_relative_error_pct_of_max': ["mean","median",p25,p75] }).\
    rename(columns={"<lambda_0>":'25%ile',"<lambda_1>":'75%ile'})
yearly_perf_summary_fullyears.columns = ['_'.join(col) for col in yearly_perf_summary_fullyears.columns.values]
yearly_perf_summary_fullyears.reset_index(inplace=True)
puddle = yearly_perf_summary_fullyears[['model','wind_source','obsgroup','relative_error_pct_of_max_median','bergey_relative_error_pct_of_max_median']]\
    .rename(columns={"relative_error_pct_of_max_median":"estimate","bergey_relative_error_pct_of_max_median":"bergey"})\
    .melt(id_vars=['model','wind_source','obsgroup'])

fig = px.bar(puddle,y="value",x="model",color="variable",barmode="group",facet_col="obsgroup",facet_row="wind_source")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

#### Performance Drill Down: ANL

fig = px.box(bysitebymonth.where(bysitebymonth.model == "anl").dropna(),x="month",y="relative_error_pct_of_max",
        facet_col="wind_source",facet_row="obsgroup",
        labels={"relative_error_pct_of_max":"Error (%)","month":"Month"},
        title="Median performance of ANL Model vs. Month of Year (all sites)")
fig.for_each_annotation(lambda a: a.update(text=a.text.split("=")[-1]))
fig.show()

dc = merged_df[(merged_df["model"] == "anl") & (merged_df["wind_source"] == "wtk") & (merged_df['obsgroup'] == "bldgsonly")].groupby(["tid","month"]).agg({"relative_error_pct_of_max": "median"}).reset_index()
px.imshow(dc.pivot(index="month",columns="tid")["relative_error_pct_of_max"],
          title="Median Error by Site and Month for ANL/WTK",
          color_continuous_scale='Portland')

dc = merged_df[(merged_df["model"] == "anl") & (merged_df["wind_source"] == "wtk_bc") & (merged_df['obsgroup'] == "bldgsonly")].groupby(["tid","month"]).agg({"relative_error_pct_of_max": "median"}).reset_index()
px.imshow(dc.pivot(index="month",columns="tid")["relative_error_pct_of_max"],
          title="Median Error by Site and Month for ANL/WTK-Bias",
          color_continuous_scale='Portland')

#### Performance Drill Down: Perera

px.box(bysitebymonth.where(bysitebymonth.model == "perera").dropna(),x="month",y="relative_error_pct_of_max",facet_col="wind_source",facet_col_wrap=3,
        labels={"relative_error_pct":"Median Relative Error (%)","month":"Month"},
        title="Median performance of Perera Model vs. Month of Year (all sites)")

dc = merged_df[(merged_df["model"] == "perera") & (merged_df["wind_source"] == "wtk")].groupby(["tid","month"]).agg({"relative_error_pct_of_max": "median"}).reset_index()
px.imshow(dc.pivot(index="month",columns="tid")["relative_error_pct_of_max"],
          title="Median Error by Site and Month for Perera/WTK",
          color_continuous_scale='Portland')

dc = merged_df[(merged_df["model"] == "perera") & (merged_df["wind_source"] == "wtk_bc")].groupby(["tid","month"]).agg({"relative_error_pct_of_max": "median"}).reset_index()
px.imshow(dc.pivot(index="month",columns="tid")["relative_error_pct_of_max"],
          title="Median Error by Site and Month for Perera/WTK-Bias",
          color_continuous_scale='Portland')

#### Performance Drill Down: No Model

px.box(bysitebymonth.where(bysitebymonth.model == "none").dropna(),x="month",y="relative_error_pct_of_max",facet_col="wind_source",facet_col_wrap=3,
        labels={"relative_error_pct":"Median Relative Error (%)","month":"Month"},
        title="Median performance of No Model vs. Month of Year (all sites)")

dc = merged_df[(merged_df["model"] == "none") & (merged_df["wind_source"] == "wtk")].groupby(["tid","month"]).agg({"relative_error_pct_of_max": "median"}).reset_index()
px.imshow(dc.pivot(index="month",columns="tid")["relative_error_pct_of_max"],
          title="Median Error by Site and Month for None/WTK",
          color_continuous_scale='Portland')

dc = merged_df[(merged_df["model"] == "none") & (merged_df["wind_source"] == "wtk_bc")].groupby(["tid","month"]).agg({"relative_error_pct_of_max": "median"}).reset_index()
px.imshow(dc.pivot(index="month",columns="tid")["relative_error_pct_of_max"],
          title="Median Error by Site and Month for None/WTK-Bias",
          color_continuous_scale='Portland')

#### Summary

 - All models produce high errors sometimes, but the bulk of errors (IQR) are within +/- 10%
 - WTK LED 2018 has very high bias, 2019 appears better
 - Bias correction makes a very large difference, especially for WTK LED
 - The Perera suite of models produces extremely high estimates in some cases, which is likely a bug
 - Overall Perera outperforms ANL (!)
 - When resource is not bias corrected, the resource data error dominates, when corrected, the obstacle model error dominates. Residual errors in the resource tend to propagate through the models.
 - Sites 34 and 183 have a high residual bias (resource is overestimated), sites 170 and 28 have low residual bias (resource is underestimated). Curiously, ANL manages to "fix" 183, but not 34. Perera fixes neither. For underestimates, the models seem ill-equipped to fix the problem.
   - Site 34: on southern coast of lake ontario, many trees (wind from SW and SE)
   - Site 183: by lakeland college, large building to the NE, otherwise open (wind primarily from S, sometimes NW)
   - Site 170: jackson college, built environment to the S (wind varies, predominately SW)
   - Site 28: stagecoach rd, rural/suburban but many mature trees around turbine (wind mostly W and NE)
