==> notebooks/validation/12 Data Statistics.ipynb <==
## Data Statistics

Caleb Phillips (caleb.phillips@nrel.gov)

Some statistics on our datasets for the data table in the paper

from common import *
import pandas as pd
import site_index

## Bergey

index = site_index.SiteIndex()
len(index.index)

daily_bergey = pd.read_csv("01 Bergey Turbine Data/daily_summaries.csv.bz2")
len(daily_bergey)

bergey_ts = pd.read_csv("01 Bergey Turbine Data/prepared_and_combined.csv.bz2")
len(bergey_ts)

daily_bergey['date'].min()

daily_bergey['date'].max()

## 3D Buildings

import fiona
obstacle_data_dir = "01 Bergey Turbine Data/3dbuildings_geojson"
n_buildings = 0
n_vertices = 0
all_buildings = []
for tid in index.tids():
    buildings = fiona.open("%s/%s.json" % (obstacle_data_dir,tid))
    n_buildings = n_buildings + len(buildings)
    for b in buildings:
        n_vertices = n_vertices + np.sum([len(x) for x in b['geometry']['coordinates']])
    all_buildings.append(buildings)
    
n_buildings

n_vertices

annotated_buildings = fiona.open("%s/all_obstacles.json" % (obstacle_data_dir))
len(annotated_buildings) - n_buildings

np.sum([len(x) for x in annotated_buildings[0]['geometry']['coordinates']])

## Met Tower Data

n_bc_sites = 0
n_bc_datapoints = 0
min_time = None
max_time = None

for tid in index.tids():
    minfo = index.lookup_by_tid(tid)
    mfile = minfo["Met Tower"]
    try:
        mdf = pd.read_csv("02 Bias Correction/%s" % mfile)
        n_bc_sites += 1
        n_bc_datapoints += len(mdf)
        dt = pd.to_datetime(mdf['Time'])
        if (min_time is None) or (dt.min() < min_time):
            min_time = dt.min()
        if (max_time is None) or (dt.max() > max_time):
            max_time = dt.max()
    except:
        print("Skip")

n_bc_sites

n_bc_datapoints

[min_time,max_time]

## WTK

wtk = pd.read_csv("01 Bergey Turbine Data/wtk.csv.bz2")
len(wtk)

wtk_led_2018 = pd.read_csv("01 Bergey Turbine Data/wtk_led_2018.csv.bz2")
wtk_led_2019 = pd.read_csv("01 Bergey Turbine Data/wtk_led_2019.csv.bz2")
len(wtk_led_2018) + len(wtk_led_2019)

wtk_bc = pd.read_csv("02 Bias Correction/wtk_met.csv.bz2")
len(wtk_bc)

wtk_led2018_bc = pd.read_csv("02 Bias Correction/wtk_led_2018_met.csv.bz2")
wtk_led2019_bc = pd.read_csv("02 Bias Correction/wtk_led_2019_met.csv.bz2")
len(wtk_led2018_bc) + len(wtk_led2019_bc)

[wtk['datetime'].min(),wtk['datetime'].max()]
